refactor: [Task 1.4] Extract TeletextPageConfig to dedicated module
